change q4

+++ src/main/java/cs455/aqi/q4/Q4Job.java
@@ -4,7 +4,7 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.Text;
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -30,7 +30,7 @@ public class Q4Job{
         job.setMapOutputValueClass(IntWritable.class);
         // Outputs from the Reducer
         job.setOutputKeyClass(Text.class);
-        job.setOutputKeyClass(LongWritable.class);
+        job.setOutputKeyClass(DoubleWritable.class);
 
         job.setNumReduceTasks(1); 
         //System.out.println(args[0]);
